@@ -1,39 +1,59 @@
-import 'migration.dart';
+import 'package:count_me_down/database/migrations.dart' as migrations;
+import 'package:count_me_down/database/migrations/v1.dart' as v1;
+import 'package:count_me_down/database/migrations/v2.dart' as v2;
 
-List<Migration> migrations = [
-  Migration(
-    version: 1,
-    actions: <String>[
-      """CREATE TABLE sessions (
-          id INTEGER PRIMARY KEY,
-          profile_id INTEGER REFERENCES profiles(id) ON DELETE CASCADE,
-          name TEXT,
-          started_at INTEGER,
-          ended_at INTEGER
-        );""",
-      """CREATE TABLE profiles (
-          id INTEGER PRIMARY KEY,
-          name TEXT,
-          body_weight INTEGER,
-          body_water_percentage INTEGER,
-          absorption_time INTEGER,
-          per_mil_metabolized_per_hour REAL
-        );""",
-      """CREATE TABLE drinks (
-          id INTEGER PRIMARY KEY,
-          session_id INTEGER REFERENCES sessions(id) ON DELETE CASCADE,
-          name TEXT,
-          volume INTEGER,
-          alcohol_concentration REAL,
-          timestamp INTEGER,
-          color INTEGER,
-          drink_type TEXT
-        );""",
-    ],
-    rollback: [
-      'DROP TABLE drinks;',
-      'DROP TABLE profiles;',
-      'DROP TABLE sessions;',
-    ],
-  )
-];
+class Migrations {
+  late final v1.Migration migrationV1;
+  late final v2.Migration migrationV2;
+
+  Migrations() {
+    migrationV1 = v1.Migration();
+    migrationV2 = v2.Migration(migrationV1: migrationV1);
+  }
+
+  List<migrations.SqfMigration> getSqfMigrations() {
+    return <migrations.SqfMigration>[
+      migrations.SqfMigration(
+        version: 1,
+        migration: migrationV1.migrate,
+        rollback: migrationV1.rollback,
+      ),
+      migrations.SqfMigration(
+        version: 2,
+        migration: migrationV2.migrate,
+        rollback: migrationV2.rollback,
+      )
+    ];
+  }
+
+  List<migrations.IdbMigration> getIdbMigrations() {
+    return <migrations.IdbMigration>[
+      migrations.IdbMigration(
+        version: 1,
+        properties: <migrations.IdbMigrationTable>[
+          migrations.IdbMigrationTable(
+            migrationV1.profileTable,
+            autoIncrement: true,
+          ),
+          migrations.IdbMigrationTable(
+            migrationV1.sessionTable,
+            autoIncrement: true,
+          ),
+          migrations.IdbMigrationTable(
+            migrationV1.drinkTable,
+            autoIncrement: true,
+          ),
+        ],
+      ),
+      migrations.IdbMigration(
+        version: 2,
+        properties: <migrations.IdbMigrationTable>[
+          migrations.IdbMigrationTable(
+            migrationV2.preferencesTable,
+            autoIncrement: true,
+          ),
+        ],
+      ),
+    ];
+  }
+}
