@@ -1,7 +1,6 @@
-import 'dart:ui';
-
 import 'package:count_me_down/extensions.dart';
 import 'package:count_me_down/models/session.dart';
+import 'package:count_me_down/utils/data_parser.dart';
 import 'package:count_me_down/utils/percent.dart';
 import 'package:count_me_down/utils/volume.dart';
 import 'package:flutter/material.dart';
@@ -9,7 +8,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
 
 class Drink {
-  static const tableName = 'drinks';
+  static const tableName = 'drink';
   static const colId = 'id';
   static const colSessionId = 'session_id';
   static const colName = 'name';
@@ -32,16 +31,16 @@ class Drink {
     'wine_glass',
   ];
   static const alcoholDensity = 0.8;
-  int id;
-  int sessionId;
-  String name;
-  Volume volume;
-  Percent alcoholConcentration;
-  DateTime timestamp;
-  Color color;
-  DrinkTypes drinkType;
+  int? id;
+  int? sessionId;
+  String? name;
+  Volume? volume;
+  Percent? alcoholConcentration;
+  DateTime? timestamp;
+  Color? color;
+  DrinkTypes? drinkType;
 
-  Session session;
+  Session? session;
 
   Drink({
     this.sessionId,
@@ -54,20 +53,16 @@ class Drink {
   });
 
   Drink.fromMap(Map<String, dynamic> map) {
-    id = map[colId];
-    sessionId = map[colSessionId];
-    name = map[colName];
+    final DataParser p = DataParser();
+
+    id = p.tryParseInt(map[colId]);
+    sessionId = p.tryParseInt(map[colSessionId]);
+    name = p.tryParseString(map[colName]);
     volume = map[colVolume] != null ? Volume(map[colVolume]) : null;
     alcoholConcentration = map[colAlcoholConcentration] != null
         ? Percent(map[colAlcoholConcentration])
         : null;
-    if (map[colTimestamp] != null) {
-      timestamp = int.tryParse(map[colTimestamp].toString()) != null
-          ? DateTime.fromMillisecondsSinceEpoch(
-              int.tryParse(map[colTimestamp].toString()),
-              isUtc: true)
-          : DateTime.tryParse(map[colTimestamp].toString());
-    }
+    timestamp = p.tryParseDateTime(map[colTimestamp]);
     color = map[colColor] != null ? Color(map[colColor]) : null;
     drinkType =
         map[colDrinkType] != null && _drinkTypes.contains(map[colDrinkType])
@@ -92,65 +87,70 @@ class Drink {
   IconData get iconData {
     switch (drinkType) {
       case DrinkTypes.beer:
-        return FontAwesomeIcons.beer;
+        return FontAwesomeIcons.beerMugEmpty;
       case DrinkTypes.blender:
         return FontAwesomeIcons.blender;
       case DrinkTypes.cocktail:
-        return FontAwesomeIcons.cocktail;
+        return FontAwesomeIcons.martiniGlassCitrus;
       case DrinkTypes.coffee:
-        return FontAwesomeIcons.coffee;
+        return FontAwesomeIcons.mugSaucer;
       case DrinkTypes.flask:
         return FontAwesomeIcons.flask;
       case DrinkTypes.glass_martini:
-        return FontAwesomeIcons.glassMartiniAlt;
+        return FontAwesomeIcons.martiniGlass;
       case DrinkTypes.glass_whiskey:
-        return FontAwesomeIcons.glassWhiskey;
+        return FontAwesomeIcons.whiskeyGlass;
       case DrinkTypes.wine_bottle:
         return FontAwesomeIcons.wineBottle;
       case DrinkTypes.wine_glass:
-        return FontAwesomeIcons.wineGlassAlt;
+        return FontAwesomeIcons.wineGlassEmpty;
       default:
-        return FontAwesomeIcons.glassWhiskey;
+        return FontAwesomeIcons.whiskeyGlass;
     }
   }
 
   Map<String, dynamic> toMap({bool forQuery = false}) {
-    final Map<String, dynamic> map = <String, dynamic>{};
+    final DataParser p = DataParser(forQuery: forQuery);
+    final DrinkTypes? drinkType = this.drinkType;
 
-    map[colId] = id;
-    map[colSessionId] = sessionId;
-    map[colName] = name;
-    map[colVolume] = volume != null ? volume.millilitres : null;
-    map[colAlcoholConcentration] =
-        alcoholConcentration != null ? alcoholConcentration.fraction : null;
-    map[colTimestamp] = timestamp != null
-        ? forQuery
-            ? timestamp.millisecondsSinceEpoch
-            : timestamp.toIso8601String()
-        : null;
-    map[colColor] = color != null ? color.value : null;
-    map[colDrinkType] = drinkType != null ? _drinkTypes[drinkType.index] : null;
+    final Map<String, dynamic> map = <String, dynamic>{
+      colId: p.serializeInt(id),
+      colSessionId: p.serializeInt(sessionId),
+      colName: p.serializeString(name),
+      colVolume: volume?.millilitres,
+      colAlcoholConcentration: alcoholConcentration?.fraction,
+      colTimestamp: p.tryParseDateTime(timestamp),
+      colColor: color?.value,
+      colDrinkType: drinkType != null ? _drinkTypes[drinkType.index] : null,
+    };
 
     if (!forQuery) {
-      map[relSession] =
-          session != null ? session.toMap(forQuery: forQuery) : null;
+      map.putIfAbsent(relSession, () => session?.toMap(forQuery: forQuery));
     }
 
     return map;
   }
 
   double get alcoholContentInGrams {
-    return volume.millilitres * alcoholConcentration.fraction * alcoholDensity;
+    return (volume?.millilitres ?? 0) *
+        (alcoholConcentration?.fraction ?? 0) *
+        alcoholDensity;
   }
 
   bool consumedBetween(DateTime from, DateTime to) {
+    final DateTime? timestamp = this.timestamp;
+
+    if (timestamp == null) return false;
+
     return (timestamp.isAfter(from) || timestamp.isAtSameMomentAs(from)) &&
         (timestamp.isBefore(to) || timestamp.isAtSameMomentAs(to));
   }
 
   double currentlyAbsorbedAlcohol(Duration absorptionTime) {
     final DateTime now = MockableDateTime.current;
-    final DateTime fullyAbsorbed = timestamp.add(absorptionTime);
+
+    final DateTime fullyAbsorbed =
+        (timestamp ?? DateTime.now()).add(absorptionTime);
     final bool alreadyAbsorbed = fullyAbsorbed.isBefore(now);
 
     if (alreadyAbsorbed) {
@@ -158,7 +158,8 @@ class Drink {
     }
 
     final Duration timeUntilAbsorption = fullyAbsorbed.difference(now);
-    final Duration totalAbsorptionTime = fullyAbsorbed.difference(timestamp);
+    final Duration totalAbsorptionTime =
+        fullyAbsorbed.difference(timestamp ?? DateTime.now());
     final bool absorptionCompleted = timeUntilAbsorption == Duration.zero;
     final bool absorptionNotStarted = totalAbsorptionTime < timeUntilAbsorption;
 
@@ -180,6 +181,12 @@ class Drink {
 
   @override
   String toString() {
+    final DateTime? timestamp = this.timestamp;
+
+    if (timestamp == null) {
+      return '(${volume.toString()}) - $name';
+    }
+
     return '${DateFormat('HH:mm').format(timestamp.toLocal())} '
         '(${volume.toString()}) - $name';
   }
