@@ -1,3 +1,4 @@
+import 'package:count_me_down/database/repos/preferences_repo.dart';
 import 'package:count_me_down/models/preferences.dart';
 import 'package:count_me_down/pages/create_session_page.dart';
 import 'package:count_me_down/pages/profile_page.dart';
@@ -8,16 +9,10 @@ import 'package:count_me_down/pages/settings_page.dart';
 import 'package:count_me_down/pages/start_page.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  final SharedPreferences sharedPreferences =
-      await SharedPreferences.getInstance();
-  final String preferencesJson = sharedPreferences.getString('preferences');
-  final Preferences preferences = preferencesJson != null
-      ? Preferences.fromJson(preferencesJson)
-      : Preferences.initialValues();
+  final Preferences preferences = await getPreferences();
 
   runApp(Provider(
     create: (_) => preferences,
@@ -32,9 +27,9 @@ class CountMeDownApp extends StatelessWidget {
 
     return MaterialApp(
         theme: ThemeData(
-          // Tyrkisk pepper blå
+          // Hot 'n Sweet blå
           primaryColor: Color.fromRGBO(1, 12, 142, 1),
-          // Jäger grønn
+          // Jägermeister grønn
           // primaryColor: Color.fromRGBO(11, 38, 16, 1),
           visualDensity: VisualDensity.adaptivePlatformDensity,
         ),
