@@ -1,8 +1,9 @@
-import 'package:count_me_down/database/db_repo.dart';
+import 'package:count_me_down/database/db_repos.dart';
+import 'package:count_me_down/database/repos/preferences_repo.dart';
 import 'package:count_me_down/models/preferences.dart';
 import 'package:count_me_down/models/session.dart';
 import 'package:count_me_down/pages/session_page.dart';
-import 'package:count_me_down/utils/utils.dart';
+import 'package:count_me_down/utils/utils.dart' as utils;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -16,7 +17,6 @@ class CreateSessionPage extends StatefulWidget {
 class _CreateSessionPageState extends State<CreateSessionPage> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final TextEditingController _nameController = TextEditingController();
-  bool _isLoading = false;
 
   @override
   Widget build(BuildContext context) {
@@ -53,7 +53,7 @@ class _CreateSessionPageState extends State<CreateSessionPage> {
                         child: Text(
                           'Start session',
                           style: TextStyle(
-                            color: Utils.getThemeTextColor(context),
+                            color: utils.getThemeTextColor(context),
                             fontSize: 17.0,
                           ),
                         ),
@@ -71,15 +71,15 @@ class _CreateSessionPageState extends State<CreateSessionPage> {
   }
 
   Future<void> _createNewSession(BuildContext context) async {
-    if (!_formKey.currentState.validate()) {
+    if (_formKey.currentState?.validate() != true) {
       return;
     }
 
-    if (mounted) {
+    /*if (mounted) {
       setState(() {
         _isLoading = true;
       });
-    }
+    }*/
 
     final Preferences preferences = context.read<Preferences>();
 
@@ -89,17 +89,17 @@ class _CreateSessionPageState extends State<CreateSessionPage> {
       startedAt: DateTime.now(),
     );
 
-    await SessionRepo.insertSession(session);
+    await insertSession(session);
     preferences.activeSessionId = session.id;
-    await preferences.save();
+    await updatePreferences(preferences);
 
     Navigator.of(context)
         .pushNamedAndRemoveUntil(SessionPage.routeName, (route) => false);
 
-    if (mounted) {
+    /*if (mounted) {
       setState(() {
         _isLoading = false;
       });
-    }
+    }*/
   }
 }
